+++ R/raad-srtm-files.R
@@ -1,0 +1,28 @@
+#' SRTM files
+#'
+#' SRTM 90m Digital Elevation Database v4.1
+#'
+#' DOI: 0.1080/13658810601169899
+#' @references [https://cgiarcsi.community/data/srtm-90m-digital-elevation-database-v4-1/]
+#' @return tibble data frame of file names, with columns `fullname`, `x`, `y` (tiles)
+#' @importFrom dplyr arrange distinct filter mutate
+#' @importFrom rlang .data
+#' @name srtm
+#' @aliases srtm
+#' @export
+#' @importFrom stringr str_detect str_extract str_replace
+#' @examples
+#' srtm_files()
+srtm_files <- function() {
+  files <- dplyr::filter(get_raad_filenames(), stringr::str_detect(.data$file, "srtm.csi.cgiar.org"))
+  files <- dplyr::filter(files, stringr::str_detect(.data$file, ".*\\.tif$"))
+  files <- dplyr::filter(files, stringr::str_detect(.data$file, "srtm_5x5/TIFF/srtm"))
+  files <-   dplyr::transmute(files, fullname = file.path(.data$root, .data$file), .data$root)
+
+  if (nrow(files) < 1)
+    stop("no smap files found")
+  ## datadir <- get_raad_datadir()
+  tile <- do.call(rbind, lapply(strsplit(basename(files$fullname), "_"), function(bs) as.integer(gsub("\\.tif$", "", bs[c(2, 3)]))))
+  files[c("x", "y")] <- tile
+  dplyr::select(files,  .data$fullname, .data$x, .data$y, .data$root)
+}